reorganizacao do codigo
