+++ index.js
@@ -4,9 +4,9 @@ const express = require('express');
 const app = express();
 
 const main = require('./src/db/connectDB.js');
-const router = require('./src/routers/routers.js');
+const AppRoutes = require('./src/routers/index.routes.js');
 
 app.use(express.json());
-app.use('/', router);
+AppRoutes(app);
 main();
 const server = app.listen(process.env.PORT, () => console.log(`Servidor online!`));
