+++ package.json
@@ -18,6 +18,8 @@
     "nodemon": "^3.0.3"
   },
   "dependencies": {
-    "dotenv": "^16.4.5"
+    "ajv": "^8.12.0",
+    "dotenv": "^16.4.5",
+    "uuid": "^9.0.1"
   }
 }
